requirements.txt
selenium==4.15.2
beautifulsoup4==4.12.2
pymongo==4.6.0
python-dotenv==1.0.0
webdriver-manager==4.0.1
lxml==4.9.3
streamlit==1.29.0
pandas==2.1.4
plotly==5.18.0
numpy==1.26.2
openai==1.106.1